+++ react/components/shared/DataEntry.tsx
@@ -1,10 +1,16 @@
 import React, { FunctionComponent } from 'react'
 
+import className from '../../styles/ContentBox.css'
+
 const DataEntry: FunctionComponent<Props> = ({ label, children }) => {
   return (
     <>
-      <label className="db c-on-base mb3">{label}</label>
-      <div className="light c-on-disabled">{children}</div>
+      <label className={`db c-on-base mb3 ${className.dataEntryLabel}`}>
+        {label}
+      </label>
+      <div className={`light c-on-disabled ${className.dataEntryChildren}`}>
+        {children}
+      </div>
     </>
   )
 }