+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+-  Corporate section data css handles
+-  Data entry custom css handles.
+
 ## [1.15.0] - 2020-06-24
 
 ### Added